Split MediaScannerService
 - Split MediaScannerService into several classes for better
   maintainability
