+++ jpsonic-main/src/test/java/com/tesshu/jpsonic/service/scanner/MediaScannerServiceImplTest.java
@@ -19,10 +19,9 @@
  * (C) 2018 tesshucom
  */
 
-package com.tesshu.jpsonic.service;
+package com.tesshu.jpsonic.service.scanner;
 
 import static com.tesshu.jpsonic.service.ServiceMockUtils.mock;
-import static com.tesshu.jpsonic.util.PlayerUtils.FAR_PAST;
 import static com.tesshu.jpsonic.util.PlayerUtils.now;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
@@ -36,14 +35,11 @@ import java.io.InputStream;
 import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.time.Instant;
-import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
@@ -61,14 +57,18 @@ import com.tesshu.jpsonic.dao.MediaFileDao;
 import com.tesshu.jpsonic.dao.MusicFolderDao;
 import com.tesshu.jpsonic.domain.Album;
 import com.tesshu.jpsonic.domain.Artist;
-import com.tesshu.jpsonic.domain.Genres;
 import com.tesshu.jpsonic.domain.MediaFile;
 import com.tesshu.jpsonic.domain.MediaFile.MediaType;
 import com.tesshu.jpsonic.domain.MediaLibraryStatistics;
 import com.tesshu.jpsonic.domain.MusicFolder;
 import com.tesshu.jpsonic.domain.SearchResult;
-import com.tesshu.jpsonic.service.metadata.MetaDataParserFactory;
-import com.tesshu.jpsonic.service.scanner.MediaScannerServiceUtils;
+import com.tesshu.jpsonic.service.MediaFileCache;
+import com.tesshu.jpsonic.service.MediaFileService;
+import com.tesshu.jpsonic.service.MediaScannerService;
+import com.tesshu.jpsonic.service.MusicFolderService;
+import com.tesshu.jpsonic.service.PlaylistService;
+import com.tesshu.jpsonic.service.SearchService;
+import com.tesshu.jpsonic.service.SettingsService;
 import com.tesshu.jpsonic.service.search.IndexManager;
 import com.tesshu.jpsonic.service.search.IndexType;
 import com.tesshu.jpsonic.service.search.SearchCriteriaDirector;
@@ -93,9 +93,9 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.util.ObjectUtils;
 
 @SuppressWarnings({ "PMD.TooManyStaticImports", "PMD.AvoidDuplicateLiterals" })
-class MediaScannerServiceTest {
+class MediaScannerServiceImplTest {
 
-    private static final Logger LOG = LoggerFactory.getLogger(MediaScannerServiceTest.class);
+    private static final Logger LOG = LoggerFactory.getLogger(MediaScannerServiceImplTest.class);
 
     @Nested
     class UnitTest {
@@ -106,8 +106,11 @@ class MediaScannerServiceTest {
         private AlbumDao albumDao;
         private MediaFileService mediaFileService;
         private MediaFileDao mediaFileDao;
-        private MediaScannerServiceImpl mediaScannerService;
+        private ScannerStateService scannerStateService;
         private ThreadPoolTaskExecutor executor;
+        private SortProcedureService utils;
+        private ScannerProcedureService scannerProcedureService;
+        private MediaScannerServiceImpl mediaScannerService;
 
         @BeforeEach
         public void setup() {
@@ -118,19 +121,15 @@ class MediaScannerServiceTest {
             artistDao = mock(ArtistDao.class);
             albumDao = mock(AlbumDao.class);
             executor = mock(ThreadPoolTaskExecutor.class);
+            utils = mock(SortProcedureService.class);
+
+            scannerStateService = new ScannerStateService(indexManager);
+            scannerProcedureService = new ScannerProcedureService(settingsService, indexManager, mediaFileService,
+                    mediaFileDao, artistDao, albumDao, utils, scannerStateService, mock(Ehcache.class),
+                    mock(MediaFileCache.class));
             mediaScannerService = new MediaScannerServiceImpl(settingsService, mock(MusicFolderService.class),
-                    indexManager, mock(PlaylistService.class), mock(MediaFileCache.class), mediaFileService,
-                    mediaFileDao, artistDao, albumDao, mock(Ehcache.class), mock(MediaScannerServiceUtils.class),
-                    executor);
-        }
-
-        @Test
-        void testNeverScanned() {
-            Mockito.when(indexManager.getStatistics()).thenReturn(null);
-            assertTrue(mediaScannerService.neverScanned());
-
-            Mockito.when(indexManager.getStatistics()).thenReturn(new MediaLibraryStatistics());
-            assertFalse(mediaScannerService.neverScanned());
+                    indexManager, null, mediaFileService, mediaFileDao, artistDao, albumDao, executor,
+                    scannerStateService, scannerProcedureService, mock(ExpungeService.class));
         }
 
         @SuppressWarnings("PMD.JUnitTestsShouldIncludeAssert") // It doesn't seem to be able to capture
@@ -143,309 +142,17 @@ class MediaScannerServiceTest {
             executor.initialize();
             indexManager = mock(IndexManager.class);
             Mockito.doNothing().when(indexManager).startIndexing();
-            Path podcastPath = Path.of(MediaScannerServiceTest.class.getResource("/MEDIAS/Scan/Null").toURI());
+            Path podcastPath = Path.of(MediaScannerServiceImplTest.class.getResource("/MEDIAS/Scan/Null").toURI());
             Mockito.when(settingsService.getPodcastFolder()).thenReturn(podcastPath.toString());
             MediaFile mediaFile = new MediaFile();
             mediaFile.setPathString(podcastPath.toString());
             Mockito.when(mediaFileService.getMediaFile(podcastPath)).thenReturn(mediaFile);
 
             mediaScannerService = new MediaScannerServiceImpl(settingsService, mock(MusicFolderService.class),
-                    indexManager, mock(PlaylistService.class), mock(MediaFileCache.class), mediaFileService,
-                    mediaFileDao, artistDao, albumDao, mock(Ehcache.class), mock(MediaScannerServiceUtils.class),
-                    executor);
+                    indexManager, mock(PlaylistService.class), mediaFileService, mediaFileDao, artistDao, albumDao,
+                    executor, scannerStateService, scannerProcedureService, mock(ExpungeService.class));
             mediaScannerService.scanLibrary();
             executor.shutdown();
-        }
-
-        @Nested
-        class ScanFileTest {
-
-            private Path createPath(String path) throws URISyntaxException {
-                return Path.of(MediaFileServiceTest.class.getResource(path).toURI());
-            }
-
-            @Test
-            void testScanFile() throws URISyntaxException, ExecutionException {
-
-                SecurityService securityService = mock(SecurityService.class);
-                MetaDataParserFactory metaDataParserFactory = mock(MetaDataParserFactory.class);
-                mediaFileService = new MediaFileService(settingsService, mock(MusicFolderService.class),
-                        securityService, mock(MediaFileCache.class), mediaFileDao, mock(AlbumDao.class),
-                        metaDataParserFactory, mock(MediaFileServiceUtils.class));
-                mediaScannerService = new MediaScannerServiceImpl(settingsService, mock(MusicFolderService.class),
-                        indexManager, mock(PlaylistService.class), mock(MediaFileCache.class), mediaFileService,
-                        mediaFileDao, artistDao, albumDao, mock(Ehcache.class), mock(MediaScannerServiceUtils.class),
-                        mock(ThreadPoolTaskExecutor.class));
-
-                assertTrue(mediaFileService.isSchemeLastModified());
-
-                Mockito.when(settingsService.getVideoFileTypesAsArray()).thenReturn(new String[0]);
-                Mockito.when(settingsService.getMusicFileTypesAsArray()).thenReturn(new String[] { "mp3" });
-                Mockito.when(securityService.isReadAllowed(Mockito.any(Path.class))).thenReturn(true);
-
-                Path dir = createPath("/MEDIAS/Music2/_DIR_ chrome hoof - 2004");
-                assertTrue(Files.isDirectory(dir));
-                MediaFile album = mediaFileService.createMediaFile(dir);
-
-                Path file = createPath("/MEDIAS/Music2/_DIR_ chrome hoof - 2004/10 telegraph hill.mp3");
-                assertFalse(Files.isDirectory(file));
-                MediaFile child = mediaFileService.createMediaFile(file);
-                child.setLastScanned(FAR_PAST);
-
-                MusicFolder musicFolder = new MusicFolder(
-                        MusicFolderTestDataUtils.resolveBaseMediaPath().concat("Music2"), "name", false, now());
-                Instant scanStart = now();
-                MediaLibraryStatistics statistics = new MediaLibraryStatistics(scanStart);
-                Map<String, Integer> albumCount = new ConcurrentHashMap<>();
-                Genres genres = new Genres();
-
-                List<MediaFile> children = Arrays.asList(child);
-                Mockito.when(mediaFileDao.getChildrenOf(album.getPathString())).thenReturn(children);
-
-                mediaScannerService.scanFile(album, musicFolder, statistics, albumCount, genres, false);
-            }
-        }
-
-        @Nested
-        class UpdateAlbumTest {
-
-            private MediaFile createSong() {
-                // nonull
-                MediaFile song = new MediaFile();
-                song.setAlbumName("albumName");
-                song.setParentPathString("parentPath");
-                song.setMediaType(MediaType.MUSIC);
-                song.setArtist("artist");
-
-                // nullable
-                song.setLastScanned(now().minus(1, ChronoUnit.SECONDS));
-                return song;
-            }
-
-            private MediaLibraryStatistics createStatistics() {
-                return new MediaLibraryStatistics(now());
-            }
-
-            private MusicFolder createMusicFolder() {
-                return new MusicFolder(Integer.valueOf(1), "", "", true, now());
-            }
-
-            @Test
-            void testIsNotAlbumUpdatable() {
-
-                MusicFolder musicFolder = createMusicFolder();
-                MediaLibraryStatistics statistics = createStatistics();
-                Map<String, Integer> albumCount = new ConcurrentHashMap<>();
-
-                MediaFile song = createSong();
-                song.setAlbumName(null);
-                mediaScannerService.updateAlbum(song, musicFolder, statistics.getScanDate(), albumCount);
-                Mockito.verify(albumDao, Mockito.never()).createOrUpdateAlbum(Mockito.any(Album.class));
-
-                song = createSong();
-                song.setParentPathString(null);
-                mediaScannerService.updateAlbum(song, musicFolder, statistics.getScanDate(), albumCount);
-                Mockito.verify(albumDao, Mockito.never()).createOrUpdateAlbum(Mockito.any(Album.class));
-
-                song = createSong();
-                song.setMediaType(MediaType.DIRECTORY);
-                mediaScannerService.updateAlbum(song, musicFolder, statistics.getScanDate(), albumCount);
-                Mockito.verify(albumDao, Mockito.never()).createOrUpdateAlbum(Mockito.any(Album.class));
-
-                song = createSong();
-                song.setAlbumArtist(null);
-                song.setArtist(null);
-                mediaScannerService.updateAlbum(song, musicFolder, statistics.getScanDate(), albumCount);
-                Mockito.verify(albumDao, Mockito.never()).createOrUpdateAlbum(Mockito.any(Album.class));
-
-                song = createSong();
-                song.setAlbumArtist("albumArtist");
-                song.setArtist(null);
-                mediaScannerService.updateAlbum(song, musicFolder, statistics.getScanDate(), albumCount);
-                Mockito.verify(albumDao, Mockito.times(1)).createOrUpdateAlbum(Mockito.any(Album.class));
-
-                Mockito.clearInvocations(albumDao);
-                song = createSong();
-                song.setAlbumArtist(null);
-                song.setArtist("artist");
-                mediaScannerService.updateAlbum(song, musicFolder, statistics.getScanDate(), albumCount);
-                Mockito.verify(albumDao, Mockito.times(1)).createOrUpdateAlbum(Mockito.any(Album.class));
-
-                Mockito.clearInvocations(albumDao);
-                song = createSong();
-                song.setAlbumArtist("albumArtist");
-                song.setArtist("artist");
-                mediaScannerService.updateAlbum(song, musicFolder, statistics.getScanDate(), albumCount);
-                Mockito.verify(albumDao, Mockito.times(1)).createOrUpdateAlbum(Mockito.any(Album.class));
-            }
-
-            @Test
-            void testFirstEncounter() {
-
-                /*
-                 * Album property determines by date whether it is the first child. In other words, with this method, it
-                 * is not possible to parallelize all child scans. (The first child must complete the scan first)
-                 */
-
-                final MediaFile song = createSong();
-                MusicFolder musicFolder = createMusicFolder();
-                MediaLibraryStatistics statistics = createStatistics();
-                Map<String, Integer> albumCount = new ConcurrentHashMap<>();
-
-                // Song dates are never updated
-                assertNotEquals(song.getLastScanned(), statistics.getScanDate());
-
-                // ## First run
-                mediaScannerService.updateAlbum(song, musicFolder, statistics.getScanDate(), albumCount);
-
-                ArgumentCaptor<MediaFile> mediaCap = ArgumentCaptor.forClass(MediaFile.class);
-                ArgumentCaptor<Album> albumCap = ArgumentCaptor.forClass(Album.class);
-                Mockito.verify(albumDao, Mockito.times(1)).createOrUpdateAlbum(albumCap.capture());
-                Mockito.verify(indexManager, Mockito.times(1)).index(Mockito.any(Album.class));
-                Mockito.verify(mediaFileDao, Mockito.times(1)).createOrUpdateMediaFile(mediaCap.capture());
-
-                Album registeredAlbum = albumCap.getValue();
-                assertEquals(registeredAlbum.getLastScanned(), statistics.getScanDate());
-                Mockito.when(albumDao.getAlbumForFile(Mockito.any(MediaFile.class))).thenReturn(registeredAlbum);
-
-                // Song dates are never updated
-                MediaFile registeredMedia = mediaCap.getValue();
-                assertNotEquals(registeredMedia.getLastScanned(), statistics.getScanDate());
-
-                // ## Second run
-                mediaScannerService.updateAlbum(song, musicFolder, statistics.getScanDate(), albumCount);
-
-                // Currently always executed
-                Mockito.verify(albumDao, Mockito.times(2)).createOrUpdateAlbum(Mockito.any(Album.class));
-
-                // Not executed if already executed
-                Mockito.verify(indexManager, Mockito.times(1)).index(Mockito.any(Album.class));
-                Mockito.verify(mediaFileDao, Mockito.times(1)).createOrUpdateMediaFile(Mockito.any(MediaFile.class));
-            }
-
-            @Test
-            void testMergeOnFirstEncount() {
-
-                /*
-                 * Year / Genre will adopt the value of the first song in the album. Jpsonic specifications.
-                 */
-
-                final MediaFile song1 = createSong();
-                song1.setYear(1111);
-                song1.setGenre("Genre1");
-                MusicFolder musicFolder = createMusicFolder();
-                MediaLibraryStatistics statistics = createStatistics();
-                Map<String, Integer> albumCount = new ConcurrentHashMap<>();
-
-                // Song dates are never updated
-                assertNotEquals(song1.getLastScanned(), statistics.getScanDate());
-
-                // ## First run
-                ArgumentCaptor<Album> albumCap = ArgumentCaptor.forClass(Album.class);
-                mediaScannerService.updateAlbum(song1, musicFolder, statistics.getScanDate(), albumCount);
-                Mockito.verify(albumDao, Mockito.times(1)).createOrUpdateAlbum(albumCap.capture());
-                Album registeredAlbum = albumCap.getValue();
-                assertEquals(registeredAlbum.getLastScanned(), statistics.getScanDate());
-
-                final MediaFile song2 = createSong();
-                song2.setYear(2222);
-                song2.setGenre("Genre2");
-                Mockito.when(albumDao.getAlbumForFile(song2)).thenReturn(registeredAlbum);
-
-                // ## Second run
-                albumCap = ArgumentCaptor.forClass(Album.class);
-                mediaScannerService.updateAlbum(song2, musicFolder, statistics.getScanDate(), albumCount);
-                Mockito.verify(albumDao, Mockito.times(2)).createOrUpdateAlbum(albumCap.capture());
-                assertEquals(2, albumCap.getAllValues().size());
-                registeredAlbum = albumCap.getAllValues().get(1);
-
-                assertEquals(Integer.valueOf(1111), registeredAlbum.getYear());
-                assertEquals("Genre1", registeredAlbum.getGenre());
-            }
-
-            @Test
-            void testGetMergedAlbum() {
-
-                final MediaFile song1 = createSong();
-                song1.setCoverArtPathString("coverArtPath1");
-
-                song1.setMusicBrainzReleaseId("musicBrainzReleaseId1");
-                MusicFolder musicFolder = createMusicFolder();
-                MediaLibraryStatistics statistics = createStatistics();
-                Map<String, Integer> albumCount = new ConcurrentHashMap<>();
-
-                final MediaFile parent = createSong();
-                Mockito.when(mediaFileService.getParentOf(Mockito.any(MediaFile.class))).thenReturn(parent);
-                parent.setCoverArtPathString("parentCoverArtPath");
-
-                // ## First run
-                ArgumentCaptor<Album> albumCap = ArgumentCaptor.forClass(Album.class);
-                mediaScannerService.updateAlbum(song1, musicFolder, statistics.getScanDate(), albumCount);
-                Mockito.verify(albumDao, Mockito.times(1)).createOrUpdateAlbum(albumCap.capture());
-
-                Album registeredAlbum = albumCap.getValue();
-                assertEquals("musicBrainzReleaseId1", registeredAlbum.getMusicBrainzReleaseId());
-                assertEquals("parentCoverArtPath", registeredAlbum.getCoverArtPath());
-
-                final MediaFile song2 = createSong();
-                song2.setCoverArtPathString("coverArtPath2");
-
-                song2.setMusicBrainzReleaseId("musicBrainzReleaseId2");
-                Mockito.when(albumDao.getAlbumForFile(song2)).thenReturn(registeredAlbum);
-
-                // ## Second run
-                albumCap = ArgumentCaptor.forClass(Album.class);
-                mediaScannerService.updateAlbum(song2, musicFolder, statistics.getScanDate(), albumCount);
-                Mockito.verify(albumDao, Mockito.times(2)).createOrUpdateAlbum(albumCap.capture());
-                assertEquals(2, albumCap.getAllValues().size());
-                registeredAlbum = albumCap.getAllValues().get(1);
-
-                assertEquals("musicBrainzReleaseId2", registeredAlbum.getMusicBrainzReleaseId());
-                assertEquals("parentCoverArtPath", registeredAlbum.getCoverArtPath());
-            }
-
-            @Test
-            void testCumulativeCount() {
-
-                /*
-                 * albumDao#createOrUpdate The number of times an album is executed is greater than the number of songs.
-                 * The reason albumDao#createOrUpdateAlbum is always called is the count logic (... can be improved by
-                 * batch)
-                 */
-
-                final MediaFile song = createSong();
-                MusicFolder musicFolder = createMusicFolder();
-                MediaLibraryStatistics statistics = createStatistics();
-                Map<String, Integer> albumCount = new ConcurrentHashMap<>();
-
-                song.setDurationSeconds(60);
-
-                // ## First run
-                mediaScannerService.updateAlbum(song, musicFolder, statistics.getScanDate(), albumCount);
-                ArgumentCaptor<MediaFile> mediaCap = ArgumentCaptor.forClass(MediaFile.class);
-                ArgumentCaptor<Album> albumCap = ArgumentCaptor.forClass(Album.class);
-                Mockito.verify(albumDao, Mockito.times(1)).createOrUpdateAlbum(albumCap.capture());
-                Mockito.verify(indexManager, Mockito.times(1)).index(Mockito.any(Album.class));
-                Mockito.verify(mediaFileDao, Mockito.times(1)).createOrUpdateMediaFile(mediaCap.capture());
-
-                Album registeredAlbum = albumCap.getValue();
-                Mockito.when(albumDao.getAlbumForFile(Mockito.any(MediaFile.class))).thenReturn(registeredAlbum);
-
-                // ### First result
-                assertEquals(60, registeredAlbum.getDurationSeconds());
-                assertEquals(1, registeredAlbum.getSongCount());
-
-                // ## Second run
-                albumCap = ArgumentCaptor.forClass(Album.class);
-                Mockito.verify(albumDao, Mockito.times(1)).createOrUpdateAlbum(albumCap.capture());
-                mediaScannerService.updateAlbum(song, musicFolder, statistics.getScanDate(), albumCount);
-
-                // ### Second result
-                registeredAlbum = albumCap.getValue();
-                Mockito.when(albumDao.getAlbumForFile(Mockito.any(MediaFile.class))).thenReturn(registeredAlbum);
-                assertEquals(120, registeredAlbum.getDurationSeconds());
-                assertEquals(2, registeredAlbum.getSongCount());
-            }
         }
 
         @Nested
@@ -481,12 +188,12 @@ class MediaScannerServiceTest {
 
                 MediaFile song = createSong();
                 song.setAlbumArtist(null);
-                mediaScannerService.updateArtist(song, musicFolder, statistics.getScanDate(), albumCount);
+                scannerProcedureService.updateArtist(song, musicFolder, statistics.getScanDate(), albumCount);
                 Mockito.verify(artistDao, Mockito.never()).createOrUpdateArtist(Mockito.any(Artist.class));
 
                 song = createSong();
                 song.setMediaType(MediaType.DIRECTORY);
-                mediaScannerService.updateArtist(song, musicFolder, statistics.getScanDate(), albumCount);
+                scannerProcedureService.updateArtist(song, musicFolder, statistics.getScanDate(), albumCount);
                 Mockito.verify(artistDao, Mockito.never()).createOrUpdateArtist(Mockito.any(Artist.class));
             }
 
@@ -502,7 +209,7 @@ class MediaScannerServiceTest {
                 assertNotEquals(song.getLastScanned(), statistics.getScanDate());
 
                 // ## First run
-                mediaScannerService.updateArtist(song, musicFolder, statistics.getScanDate(), albumCount);
+                scannerProcedureService.updateArtist(song, musicFolder, statistics.getScanDate(), albumCount);
                 Mockito.verify(artistDao, Mockito.times(1)).createOrUpdateArtist(Mockito.any(Artist.class));
 
                 ArgumentCaptor<Artist> artistCap = ArgumentCaptor.forClass(Artist.class);
@@ -519,7 +226,7 @@ class MediaScannerServiceTest {
 
                 // ## Second run
                 artistCap = ArgumentCaptor.forClass(Artist.class);
-                mediaScannerService.updateArtist(song, musicFolder, statistics.getScanDate(), albumCount);
+                scannerProcedureService.updateArtist(song, musicFolder, statistics.getScanDate(), albumCount);
 
                 // Currently always executed
                 Mockito.verify(artistDao, Mockito.times(2)).createOrUpdateArtist(artistCap.capture());
@@ -812,7 +519,7 @@ class MediaScannerServiceTest {
             this.musicFolders = Arrays.asList(new MusicFolder(1, tempDir.toString(), "musicFolder", true, now()));
 
             // Copy the song file from the test resource. No tags are registered in this file.
-            Path sample = Path.of(MediaScannerServiceTest.class
+            Path sample = Path.of(MediaScannerServiceImplTest.class
                     .getResource("/MEDIAS/Scan/Timestamp/ARTIST/ALBUM/sample.mp3").toURI());
             this.song = Path.of(this.album.toString(), "sample.mp3");
             assertNotNull(Files.copy(sample, song));
@@ -839,7 +546,7 @@ class MediaScannerServiceTest {
 
             // Copy the song file from the test resource. Tags are registered in this file.
             FileUtil.deleteIfExists(this.song);
-            Path sampleEdited = Path.of(MediaScannerServiceTest.class
+            Path sampleEdited = Path.of(MediaScannerServiceImplTest.class
                     .getResource("/MEDIAS/Scan/Timestamp/ARTIST/ALBUM/sampleEdited.mp3").toURI());
             this.song = Path.of(this.album.toString(), "sample.mp3");
             Files.copy(sampleEdited, this.song);
@@ -999,7 +706,7 @@ class MediaScannerServiceTest {
             MusicFolderTestDataUtils.getTestMusicFolders().forEach(musicFolderDao::createMusicFolder);
             musicFolderService.clearMusicFolderCache();
 
-            Timer globalTimer = metrics.timer(MetricRegistry.name(MediaScannerServiceTest.class, "Timer.global"));
+            Timer globalTimer = metrics.timer(MetricRegistry.name(MediaScannerServiceImplTest.class, "Timer.global"));
 
             try (Timer.Context globalTimerContext = globalTimer.time()) {
                 TestCaseUtils.execScan(mediaScannerService);
